requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3
pycryptodome==3.21.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
itsdangerous==2.2.0
email-validator==2.2.0
bcrypt==4.0.1